--- chart/templates/crcin-deploy.yml
@@ -1,63 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: {{ include "relaynet-internet-gateway.fullname" . }}-crcin
-  labels:
-    {{- include "relaynet-internet-gateway.labels" (merge (dict "Component" "crcin") .) | nindent 4 }}
-spec:
-{{- if not .Values.autoscaling.enabled }}
-  replicas: 1
-{{- end }}
-  selector:
-    matchLabels:
-      {{- include "relaynet-internet-gateway.selectorLabels" (merge (dict "Component" "crcin") .) | nindent 6 }}
-  template:
-    metadata:
-      annotations:
-        {{- if .Values.podAnnotations }}
-        {{- toYaml .Values.podAnnotations | nindent 8 }}
-        {{- end }}
-        global-cm-digest: {{ include "relaynet-internet-gateway.resourceDigest" (merge (dict "fileName" "global-cm.yml") .) }}
-        mongo-cm-digest: {{ include "relaynet-internet-gateway.resourceDigest" (merge (dict "fileName" "mongo-cm.yml") .) }}
-        global-secret-digest: {{ include "relaynet-internet-gateway.resourceDigest" (merge (dict "fileName" "global-secret.yml") .) }}
-        keystore-cm-digest: {{ include "relaynet-internet-gateway.resourceDigest" (merge (dict "fileName" "keystore-cm.yml") .) }}
-        keystore-secret-digest: {{ include "relaynet-internet-gateway.resourceDigest" (merge (dict "fileName" "keystore-secret.yml") .) }}
-      labels:
-        {{- include "relaynet-internet-gateway.selectorLabels" (merge (dict "Component" "crcin") .) | nindent 8 }}
-    spec:
-      serviceAccountName: {{ include "relaynet-internet-gateway.serviceAccountName" . }}
-      securityContext:
-        {{- toYaml .Values.podSecurityContext | nindent 8 }}
-      shareProcessNamespace: true
-      containers:
-        - name: crcin
-          securityContext:
-            {{- toYaml .Values.securityContext | nindent 12 }}
-          image: {{ include "relaynet-internet-gateway.image" . }}
-          imagePullPolicy: IfNotPresent
-          args:
-            - build/main/bin/crc-queue-worker.js
-          env:
-            - name: WORKER_ID
-              valueFrom:
-                fieldRef:
-                  fieldPath: metadata.name
-          envFrom:
-            - configMapRef:
-                name: {{ include "relaynet-internet-gateway.fullname" . }}
-            - configMapRef:
-                name: {{ include "relaynet-internet-gateway.fullname" . }}-mongo
-            - secretRef:
-                name: {{ include "relaynet-internet-gateway.fullname" . }}
-            - configMapRef:
-                name: {{ include "relaynet-internet-gateway.fullname" . }}-keystore
-            - secretRef:
-                name: {{ include "relaynet-internet-gateway.fullname" . }}-keystore
-      {{- with .Values.nodeSelector }}
-      nodeSelector:
-        {{- toYaml . | nindent 8 }}
-      {{- end }}
-      {{- with .Values.tolerations }}
-      tolerations:
-        {{- toYaml . | nindent 8 }}
-      {{- end }}
