--- src/bin/crc-queue-worker.ts
@@ -1,6 +1,0 @@
-import { get as getEnvVar } from 'env-var';
-
-import { processIncomingCrcCargo } from '../queueWorkers/crcIncoming';
-
-const WORKER_ID = getEnvVar('WORKER_ID').required().asString();
-processIncomingCrcCargo(WORKER_ID);
