+++ src/services/queue/sinks/types.ts
@@ -1,3 +1,4 @@
 export const EVENT_TYPES = {
   PDC_OUTGOING_PARCEL: 'tech.relaycorp.awala.internet-gateway.pdc-outgoing-parcel',
+  CRC_INCOMING_CARGO: 'tech.relaycorp.awala.internet-gateway.crc-incoming-cargo',
 };
