+++ src/utilities/http.ts
@@ -10,5 +10,6 @@ export const HTTP_STATUS_CODES = {
   CONFLICT: 409,
   UNSUPPORTED_MEDIA_TYPE: 415,
   FAILED_DEPENDENCY: 424,
+  INTERNAL_SERVER_ERROR: 500,
   SERVICE_UNAVAILABLE: 503,
 } as const;
